requirements.txt
Flask==2.2.3
python-telegram-bot==20.0
requests==2.28.1
pandas==2.0.0

